# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.SendOTPRequest:
    properties:
      email:
        description: Email address to send OTP to
        type: string
    required:
    - email
    type: object
  handler.UpdateProfileRequest:
    properties:
      email:
        description: User's email address (optional update)
        type: string
      first_name:
        description: User's first name
        type: string
      last_name:
        description: User's last name
        type: string
    type: object
...
  handler.VerifyOTPRequest:
    properties:
      code:
        description: OTP code to verify
        type: string
      email:
        description: User's email address
        type: string
    required:
    - code
    - email
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@authentio.com
    name: Authentio Support
    url: https://authentio.com/support
  description: Secure authentication service with JWT, OAuth2, 2FA, and password reset
    functionality
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Authentio API
  version: "1.0"
paths:
  /2fa/disableOtp:
    post:
      consumes:
      - application/json
      description: Disable two-factor authentication for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: 2FA disabled successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Failed to disable 2FA
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - Invalid or missing JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Disable 2FA
      tags:
      - 2fa
  /2fa/enableOtp:
    post:
      consumes:
      - application/json
      description: Enable two-factor authentication using email for the authenticated
        user
      produces:
      - application/json
      responses:
        "200":
          description: 2FA enabled successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Failed to enable 2FA
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - Invalid or missing JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Enable email-based 2FA
      tags:
      - 2fa
  /2fa/sendOtp:
    post:
      consumes:
      - application/json
      description: Send a one-time password to the user's email for two-factor authentication
      parameters:
      - description: Email address to send OTP
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.SendOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP sent successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid email format or user not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to send OTP email
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Send 2FA OTP code
      tags:
      - 2fa
  /2fa/verifyOtp:
    post:
      consumes:
      - application/json
      description: Verify the one-time password for two-factor authentication
      parameters:
      - description: OTP verification data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.VerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP verified successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid OTP code, expired code, or invalid email
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify 2FA OTP code
      tags:
      - 2fa
...
  /user/getProfile:
    get:
      consumes:
      - application/json
      description: Retrieve the authenticated user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: User profile retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - Invalid or missing JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - user
  /user/updateProfile:
    put:
      consumes:
      - application/json
      description: Update the authenticated user's profile information
      parameters:
      - description: Profile update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input data
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - Invalid or missing JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Email already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - user
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {your_jwt_token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"